Revert "Redux"
