--- components/layout.tsx
@@ -1,14 +1,0 @@
-import Nav from './nav';
-
-type Props = {
-  children: any
-}
-
-const Layout = (props: Props) => (
-  <>
-    <Nav />
-    {props.children}
-  </>
-);
-
-export default Layout;
