+++ components/nav.js
@@ -2,8 +2,8 @@ import React from 'react'
 import Link from 'next/link'
 
 const links = [
-  { href: 'https://zeit.co/now', label: 'ZEIT', key: '' },
-  { href: 'https://github.com/zeit/next.js', label: 'GitHub', key: '' }
+  { href: 'https://zeit.co/now', label: 'ZEIT' },
+  { href: 'https://github.com/zeit/next.js', label: 'GitHub' }
 ].map(link => {
   link.key = `nav-link-${link.href}-${link.label}`
   return link
@@ -15,16 +15,6 @@ const Nav = () => (
       <li>
         <Link href='/'>
           <a>Home</a>
-        </Link>
-      </li>
-      <li>
-        <Link href='/search'>
-          <a>Search</a>
-        </Link>
-      </li>
-      <li>
-        <Link href='/exoplanet/test-planet'>
-          <a>Search</a>
         </Link>
       </li>
       {links.map(({ key, href, label }) => (
