--- next-env.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="next" />
-/// <reference types="next/types/global" />
