--- next.config.js
@@ -1,4 +1,0 @@
-const withSass = require('@zeit/next-sass')
-module.exports = withSass({
-  /* config options here */
-})
