+++ package.json
@@ -8,29 +8,8 @@
     "start": "next start"
   },
   "dependencies": {
-    "@types/reactstrap": "^8.0.6",
-    "@zeit/next-sass": "^1.0.1",
-    "bootstrap": "^4.3.1",
-    "isomorphic-unfetch": "^3.0.0",
-    "jquery": "^3.4.1",
     "next": "9.1.3",
-    "next-redux-wrapper": "^4.0.1",
-    "node-sass": "^4.13.0",
-    "popper.js": "^1.16.0",
     "react": "16.11.0",
-    "react-dom": "16.11.0",
-    "react-redux": "^7.1.3",
-    "reactstrap": "^8.1.1",
-    "redux": "^4.0.4",
-    "redux-devtools-extension": "^2.13.8",
-    "redux-thunk": "^2.3.0",
-    "webpack": "^4.41.2"
-  },
-  "devDependencies": {
-    "@types/next": "^9.0.0",
-    "@types/node": "^12.12.7",
-    "@types/react": "^16.9.11",
-    "@types/react-redux": "^7.1.5",
-    "typescript": "^3.7.2"
+    "react-dom": "16.11.0"
   }
 }
