--- pages/_app.tsx
@@ -1,31 +1,0 @@
-import React from "react";
-import {createStore} from "redux";
-import {Provider} from "react-redux";
-import App, { AppContext, AppInitialProps } from "next/app";
-import withRedux from "next-redux-wrapper";
-import configureStore from '../store/store';
-
-interface Props extends AppInitialProps {
-  store?: App
-}
-
-export default withRedux(configureStore)(
-  class MyApp extends App {
-    static async getInitialProps({ Component, ctx }: AppContext) {
-      return {
-        pageProps: Component.getInitialProps
-          ? await Component.getInitialProps(ctx)
-          : {},
-      }
-    }
-
-    render() {
-      const { Component, pageProps, store } = this.props as any;
-      return (
-        <Provider store={store}>
-          <Component {...pageProps} />
-        </Provider>
-      )
-    }
-  }
-)
