--- pages/posts.tsx
@@ -1,23 +1,0 @@
-import Layout from '../components/layout';
-import Link from 'next/link';
-
-const PostLink = (props: { id: string}) => (
-  <li>
-    <Link href="/p/[id]" as={`/p/asdf`}>
-      <a>{props.id}</a>
-    </Link>
-  </li>
-);
-
-export default function Blog() {
-  return (
-    <Layout>
-      <h1>My Blog</h1>
-      <ul>
-        <PostLink id="hello-nextjs" />
-        <PostLink id="learn-nextjs" />
-        <PostLink id="deploy-nextjs" />
-      </ul>
-    </Layout>
-  );
-}
