--- store/root.ts
@@ -1,10 +1,0 @@
-// https://redux.js.org/recipes/usage-with-typescript
-
-import { combineReducers } from 'redux';
-import searchReducer from './search/search.reducer';
-
-const rootReducer = combineReducers({
-    search: searchReducer,
-});
-  
-export default rootReducer;
